Adds Windows 2022 VM workload application
The application will spawn a Windows 10 VM in kubevirt. As a prerequisite,
the application will require the PVC "win2k22-qcow2" to already exist
in order to be created and started properly (the PVC can be created by
uploading an image through CDI).

The VM will also have cloudbase-init userdata, configuring the VMs username and password.